+++ lib/features/auth/data/auth_repository.dart
@@ -1,5 +1,3 @@
-import 'dart:math';
-
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:pudding/core/providers/firebase_providers.dart';