+++ lib/features/auth/presentation/auth_gate.dart
@@ -1,5 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:flutter_smart_dialog/flutter_smart_dialog.dart';
+import 'package:pudding/core/components/page_view_wrappers.dart';
+import 'package:pudding/features/auth/presentation/signin_anony.dart';
 import 'package:pudding/features/auth/providers/auth_providers.dart';
 
 class AuthGate extends ConsumerWidget {
@@ -7,14 +10,34 @@ class AuthGate extends ConsumerWidget {
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
-    return Scaffold(
+    return pageViewWrapper(
       body: Center(
-        child: ElevatedButton(
-          child: const Text('Sign In Anonymously'),
-          onPressed: () {
-            // We use ref.read() inside a callback to call a function on the provider.
-            ref.read(authRepositoryProvider).signInAnonymously();
-          },
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          spacing: 32,
+          children: [
+            ElevatedButton.icon(
+              onPressed: () => SmartDialog.show(
+                builder: (context) => SigninAnonyPanel(),
+                permanent: true,
+                alignment: Alignment.centerRight,
+                usePenetrate: true,
+                clickMaskDismiss: false,
+              ),
+              label: const Text("Sign in anonymously"),
+            ),
+            // TODO: customize
+            ElevatedButton.icon(
+              onPressed: () => SmartDialog.show(
+                builder: (context) => SigninAnonyPanel(),
+                permanent: true,
+                alignment: Alignment.centerRight,
+                usePenetrate: true,
+                clickMaskDismiss: false,
+              ),
+              label: const Text("Sign in with password"),
+            ),
+          ],
         ),
       ),
     );